Voting system removed

+++ src/command.js
@@ -98,7 +98,8 @@ function registerCommands(config) {
         }
 
         chan.getRandomPost(board).then((post) => {
-            sendPost(post, message, config);
+            let cfg = message.guild ? config.guilds.getConfigForGuild(message.guild.id) : undefined;
+            sendPost(post, message, config, cfg);
         }, (reason) => {
             if(reason.board_not_found) {
                 message.channel.send(strings["random_noboard"].format(reason.board_not_found));
@@ -116,7 +117,8 @@ function registerCommands(config) {
             let board = chan.getBoardName(args[1]);
 
             chan.getPost(id, board).then((post) => {
-                sendPost(post, message, config);
+                let cfg = message.guild ? config.guilds.getConfigForGuild(message.guild.id) : undefined;
+                sendPost(post, message, config, cfg);
             }, (reason) => {
                 if(reason.post_not_found) {
                     message.channel.send(strings["post_nopost"].format(reason.post_not_found, board));
@@ -277,7 +279,7 @@ function registerCommands(config) {
     };
 }
 
-function sendPost(post, message, config) {
+function sendPost(post, message, config, gconfig) {
     let postText = unescape(post.text.length > 2000 ? post.text.substring(0, 2000) + "..." : post.text);
     postText = postText.replace(/<br>/gi, "\n");
     postText = postText.replace("</span>", "");
@@ -288,21 +290,28 @@ function sendPost(post, message, config) {
         .setTitle(strings["post_title"].format(post.id, post.author))
         .setDescription(strings["post_desc"].format(postText, post.permalink))
         .setImage(post.image)
-        .addField(strings["post_submitted"], post.timestamp)
-        .setFooter(strings["post_removal_instructions"].format(config.removal_vote_emote, config.removal_vote_count));
-    message.channel.send(embed).then(msg => {
-        let filter = (reaction) => {
-            return reaction.emoji.name === config.removal_vote_emote;
-        };
-        msg.awaitReactions(filter, { max: config.removal_vote_count, time: 60000, errors: ["time"] }).then(collected => {
-            let reaction = collected.first();
-            if(reaction.emoji.name === config.removal_vote_emote) {
-                msg.delete().then(() => {
-                    message.channel.send(strings["post_removal_confirm"]);
-                });
-            }
-        }).catch(); // do nothing on error
-    });
+        .addField(strings["post_submitted"], post.timestamp);
+
+    let removalTime = gconfig && gconfig.removal_time ? gconfig.removal_time : 120;
+    if(gconfig) {
+        embed.setFooter(strings["post_removal_instructions"].format(config.removal_vote_emote, removalTime));
+    }
+    let msgPromise = message.channel.send(embed);
+    if(gconfig) {
+        msgPromise.then(msg => {
+            let filter = (reaction, user) => {
+                return reaction.emoji.name === config.removal_vote_emote && user.tag === message.author.tag;
+            };
+            msg.awaitReactions(filter, { max: 1, time: removalTime * 1000, errors: ["time"] }).then(collected => {
+                let reaction = collected.first();
+                if(reaction.emoji.name === config.removal_vote_emote) {
+                    msg.delete().then(() => {
+                        message.channel.send(strings["post_removal_confirm"]);
+                    });
+                }
+            }).catch(() => { /* do nothing on error */ });
+        });
+    }
 }
 
 function getPrefix(message, config) {
